+++ .idea/gradle.xml
@@ -12,6 +12,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/data" />
           </set>
         </option>
       </GradleProjectSettings>
