+++ data/src/main/java/com/whitebeach/data/remote/model/player/Shots.kt
